# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

// DefaultErrorResponse defines model for DefaultErrorResponse.
type DefaultErrorResponse struct {
	// Remarks error description
	Remarks string `json:"remarks"`
}

// TransactionRequest defines model for TransactionRequest.
type TransactionRequest struct {
	// NoRekening user bank account number
	NoRekening string `json:"no_rekening" validate:"required"`

	// Nominal amount to be stored in IDR
	Nominal int64 `json:"nominal" validate:"required,gt=0"`
}

// TransactionResponse defines model for TransactionResponse.
type TransactionResponse struct {
	// Saldo user bank account balance
	Saldo int64 `json:"saldo"`
}

// UserPostRequest defines model for UserPostRequest.
type UserPostRequest struct {
	// Nama name of the user
	Nama string `json:"nama" validate:"required"`

	// Nik user personal identification number
	Nik string `json:"nik" validate:"required,len=16"`

	// NoHp user phone number
	NoHp string `json:"no_hp" validate:"required"`
}

// UserPostResponse defines model for UserPostResponse.
type UserPostResponse struct {
	// NoRekening user bank account number
	NoRekening string `json:"no_rekening"`
}

// AccountNumberParam defines model for AccountNumberParam.
type AccountNumberParam = string

// UserPostJSONRequestBody defines body for UserPost for application/json ContentType.
type UserPostJSONRequestBody = UserPostRequest

// TransactionCreditPostJSONRequestBody defines body for TransactionCreditPost for application/json ContentType.
type TransactionCreditPostJSONRequestBody = TransactionRequest

// TransactionDebitPostJSONRequestBody defines body for TransactionDebitPost for application/json ContentType.
type TransactionDebitPostJSONRequestBody = TransactionRequest