# api/api.yaml
openapi: 3.0.0
info:
  title: Users API
  description: API to perform user data manipulation
  version: 1.0.0
...
paths:
  /daftar:
    post:
      tags:
        - users
      summary: Register a new user
      description: Register a new user along with bank account creation
      operationId: UserPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPostRequest'
      responses:
        '200':
          description: Success - Returns detail of user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPostResponse'
        '400':
          description: Error - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
...
components:
...
  schemas:
    DefaultErrorResponse:
      type: object
      required:
        - remarks
      properties:
        remarks:
          type: string
          description: error description
          example: failed to validate nama field
    UserPostRequest:
      type: object
      required:
        - nama
        - nik
        - no_hp
      properties:
        nama:
          type: string
          description: name of the user
          example: John Doe
          x-oapi-codegen-extra-tags:
            validate: required
        nik:
          type: string
          description: user personal identification number
          example: 7293847582837251
          x-oapi-codegen-extra-tags:
            validate: required,len=16
        no_hp:
          type: string
          description: user phone number
          example: 081123456789
          x-oapi-codegen-extra-tags:
            validate: required
    UserPostResponse:
      type: object
      required:
        - no_rekening
      properties:
        no_rekening:
          type: string
          description: user bank account number
          example: 2342342342
...